+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.7 (WIP)
 
 * Add `getFullYear()` method to our custom Date classes
+* Add `Number#toSmartDecimal()`
 
 ## 0.9.6 (2026-01-21)